backend/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
av==16.0.1
bcrypt==4.2.1
boto3==1.41.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
ctranslate2==4.6.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.3
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.5
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyreadline3==3.5.4
pytest==9.0.1
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
ruff==0.14.6
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
